standardize middleware (#3066)
