+++ core/metrics/instrumentation/slackertrace/doc.go
@@ -1,0 +1,2 @@
+// Package slackertrace provides a Slacker trace exporter for OpenTelemetry.
+package slackertrace
